# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Realiza a autenticação com e-mail e senha, retornando o token de acesso JWT.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "E-mail cadastrado."
                  },
                  "password": {
                    "type": "string",
                    "description": "Senha do usuário."
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "valido": {
                  "summary": "Exemplo válido",
                  "value": {
                    "email": "usuario@email.com",
                    "password": "SenhaSegura123"
                  }
                },
                "invalido": {
                  "summary": "Exemplo inválido",
                  "value": {
                    "email": "email-invalido",
                    "password": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "201": {
            "description": "Sessão criada com sucesso."
          },
          "400": {
            "description": "Dados de entrada malformados ou inválidos."
          },
          "401": {
            "description": "Credenciais incorretas."
          },
          "403": {
            "description": "Acesso recusado ou conta suspensa."
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "409": {
            "description": "Conflito na autenticação ou múltiplas sessões ativas."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        },
        "summary": "Autenticar usuário",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/matches": {
      "get": {
        "operationId": "MatchesController_findAll",
        "parameters": [
          {
            "name": "championshipId",
            "required": false,
            "in": "query",
            "description": "Filtra por campeonato",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "friendly",
            "required": false,
            "in": "query",
            "description": "true = só amistosos (sem campeonato)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownClub",
            "required": false,
            "in": "query",
            "description": "true = só partidas do clube da casa",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "upcoming",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de partidas retornada com sucesso"
          }
        },
        "summary": "Lista partidas, filtradas no servidor pelos parâmetros informados",
        "tags": [
          "matches"
        ]
      },
      "post": {
        "operationId": "MatchesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Partida cadastrada com sucesso"
          },
          "400": {
            "description": "Dados de entrada malformados ou inválidos"
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cadastra uma nova partida (Apenas ADMIN)",
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{id}": {
      "get": {
        "operationId": "MatchesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partida encontrada"
          },
          "404": {
            "description": "Partida não encontrada"
          }
        },
        "summary": "Retorna os detalhes de uma partida específica por id",
        "tags": [
          "matches"
        ]
      },
      "put": {
        "operationId": "MatchesController_replace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Partida substituída com sucesso"
          },
          "400": {
            "description": "Dados de entrada malformados ou inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Partida não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Substitui completamente os dados de uma partida (Apenas ADMIN)",
        "tags": [
          "matches"
        ]
      },
      "patch": {
        "operationId": "MatchesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partida atualizada com sucesso"
          },
          "400": {
            "description": "Dados de entrada malformados ou inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Partida não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualiza parcialmente uma partida (Apenas ADMIN)",
        "tags": [
          "matches"
        ]
      },
      "delete": {
        "operationId": "MatchesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Partida removida com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Partida não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove uma partida por id (Apenas ADMIN)",
        "tags": [
          "matches"
        ]
      }
    },
    "/ties": {
      "get": {
        "operationId": "TiesController_getBracket",
        "parameters": [
          {
            "name": "championshipId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chaveamento retornado com sucesso"
          }
        },
        "summary": "Retorna o chaveamento (confrontos) de um campeonato",
        "tags": [
          "ties"
        ]
      }
    },
    "/ties/generate": {
      "post": {
        "operationId": "TiesController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBracketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chaveamento gerado"
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gera (ou regenera) o chaveamento do mata-mata (Apenas ADMIN)",
        "tags": [
          "ties"
        ]
      }
    },
    "/ties/{id}": {
      "patch": {
        "operationId": "TiesController_updateTie",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confronto atualizado"
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Define times/placares de um confronto e recalcula o avanço (Apenas ADMIN)",
        "tags": [
          "ties"
        ]
      }
    },
...
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMatchDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Data e hora da partida no formato ISO 8601",
            "example": "2026-06-22T21:00:00.000Z"
          },
          "homeTeamId": {
            "type": "number",
            "description": "ID do time mandante. Obrigatório em partidas de campeonato.",
            "example": 1
          },
          "awayTeamId": {
            "type": "number",
            "description": "ID do time visitante. Obrigatório em partidas de campeonato.",
            "example": 2
          },
          "opponent": {
            "type": "string",
            "description": "Nome textual do adversário (legado/amistosos sem time cadastrado). Ignorado quando os times são informados por ID.",
            "example": "Tabajara FC"
          },
          "location": {
            "type": "string",
            "description": "Local onde a partida será realizada",
            "example": "Estádio Pinheirão"
          },
          "homeScore": {
            "type": "object",
            "description": "Gols do mandante. Omitir (ou null) para partida ainda não realizada.",
            "example": 3,
            "minimum": 0,
            "nullable": true
          },
          "awayScore": {
            "type": "object",
            "description": "Gols do visitante. Omitir (ou null) para partida ainda não realizada.",
            "example": 1,
            "minimum": 0,
            "nullable": true
          },
          "championshipId": {
            "type": "object",
            "description": "ID do campeonato associado (omitir para amistoso)",
            "example": 1
          },
          "homePenalties": {
            "type": "object",
            "description": "Pênaltis do mandante (preencher só quando houve disputa)",
            "example": 4,
            "minimum": 0,
            "nullable": true
          },
          "awayPenalties": {
            "type": "object",
            "description": "Pênaltis do visitante (preencher só quando houve disputa)",
            "example": 2,
            "minimum": 0,
            "nullable": true
          },
          "phase": {
            "type": "object",
            "description": "Fase do mata-mata (texto livre): \"Final\", \"Semifinal\", \"Oitavas\"...",
            "example": "Semifinal",
            "nullable": true
          },
          "groupName": {
            "type": "object",
            "description": "Grupo da fase de grupos (ex: \"A\"). Só em campeonatos com grupos.",
            "example": "A",
            "nullable": true
          }
        },
        "required": [
          "date",
          "location"
        ]
      },
      "GenerateBracketDto": {
        "type": "object",
        "properties": {
          "championshipId": {
            "type": "number",
            "description": "ID do campeonato",
            "example": 1
          },
          "teams": {
            "type": "number",
            "description": "Número de times na fase inicial (potência de 2)",
            "enum": [
              2,
              4,
              8,
              16,
              32
            ],
            "example": 8
          }
        },
        "required": [
          "championshipId",
          "teams"
        ]
      },
      "TieLegDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "object",
            "description": "Data e hora do jogo (ISO 8601)",
            "nullable": true
          },
          "location": {
            "type": "object",
            "description": "Local do jogo",
            "nullable": true
          },
          "homeScore": {
            "type": "object",
            "nullable": true
          },
          "awayScore": {
            "type": "object",
            "nullable": true
          },
          "homePenalties": {
            "type": "object",
            "nullable": true
          },
          "awayPenalties": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "UpdateTieDto": {
        "type": "object",
        "properties": {
          "homeTeamId": {
            "type": "object",
            "description": "Time mandante (só na fase inicial)",
            "nullable": true
          },
          "awayTeamId": {
            "type": "object",
            "description": "Time visitante (só na fase inicial)",
            "nullable": true
          },
          "legs": {
            "description": "Placares dos jogos do confronto (1 ou 2, conforme o campeonato)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TieLegDto"
            }
          }
        }
      },
...
    }
  }
}